+++ src/v/cluster/tests/topic_updates_dispatcher_test.cc
@@ -170,7 +170,7 @@ FIXTURE_TEST(
     auto check_allocated_counts = [&](std::vector<size_t> expected) {
         std::vector<size_t> counts;
         for (const auto& [id, node] : allocation_nodes) {
-            BOOST_REQUIRE(id() == counts.size() + 1); // 1-based node ids
+            BOOST_REQUIRE(id() == static_cast<int>(counts.size()) + 1); // 1-based node ids
             counts.push_back(node->allocated_partitions());
         }
         logger.debug("allocated counts: {}, expected: {}", counts, expected);
@@ -180,7 +180,7 @@ FIXTURE_TEST(
     auto check_final_counts = [&](std::vector<size_t> expected) {
         std::vector<size_t> counts;
         for (const auto& [id, node] : allocation_nodes) {
-            BOOST_REQUIRE(id() == counts.size() + 1); // 1-based node ids
+            BOOST_REQUIRE(id() == static_cast<int>(counts.size()) + 1); // 1-based node ids
             counts.push_back(node->final_partitions());
         }
         logger.debug("final counts: {}, expected: {}", counts, expected);
